Start migration of version fixtures
A new method of defining version fixtures has been designed.
This starts the migration towards the new format.
